refactor(react): TSify example
