+++ react/package.json
@@ -34,5 +34,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@types/jest": "~26.0.24",
+    "typescript": "~4.3.5"
   }
 }
